split hashset/hashmap into separate files
