+++ tests/z_hashset_template_test.c
@@ -25,39 +25,38 @@
 
 // ── Instantiate a uint32_t set, small initial capacity to force growth ───────
 //
-// A set is generated by leaving _ZP_HASHMAP_TEMPLATE_VAL_TYPE undefined.
 // Identity hash: bucket = key % capacity. Using the identity makes collisions
 // fully predictable, so the collision-chain tests below can deliberately place
 // several keys into the same bucket.
 static inline size_t u32_hash(const uint32_t *k) { return (size_t)(*k); }
 static inline bool u32_eq(const uint32_t *a, const uint32_t *b) { return *a == *b; }
 
-#define _ZP_HASHMAP_TEMPLATE_KEY_TYPE uint32_t
-#define _ZP_HASHMAP_TEMPLATE_NAME u32set
-#define _ZP_HASHMAP_TEMPLATE_INITIAL_CAPACITY 4
-#define _ZP_HASHMAP_TEMPLATE_KEY_HASH_FN u32_hash
-#define _ZP_HASHMAP_TEMPLATE_KEY_EQ_FN u32_eq
-#include "zenoh-pico/collections/hashmap_template.h"
+#define _ZP_HASHSET_TEMPLATE_KEY_TYPE uint32_t
+#define _ZP_HASHSET_TEMPLATE_NAME u32set
+#define _ZP_HASHSET_TEMPLATE_INITIAL_CAPACITY 4
+#define _ZP_HASHSET_TEMPLATE_KEY_HASH_FN u32_hash
+#define _ZP_HASHSET_TEMPLATE_KEY_EQ_FN u32_eq
+#include "zenoh-pico/collections/hashset_template.h"
 
 // ── Instantiate a uint8_t-indexed set to exercise the capacity limit ─────────
 
-#define _ZP_HASHMAP_TEMPLATE_KEY_TYPE uint32_t
-#define _ZP_HASHMAP_TEMPLATE_NAME u8idxset
-#define _ZP_HASHMAP_TEMPLATE_INDEX_TYPE uint8_t
-#define _ZP_HASHMAP_TEMPLATE_INITIAL_CAPACITY 2
-#define _ZP_HASHMAP_TEMPLATE_KEY_HASH_FN u32_hash
-#define _ZP_HASHMAP_TEMPLATE_KEY_EQ_FN u32_eq
-#include "zenoh-pico/collections/hashmap_template.h"
+#define _ZP_HASHSET_TEMPLATE_KEY_TYPE uint32_t
+#define _ZP_HASHSET_TEMPLATE_NAME u8idxset
+#define _ZP_HASHSET_TEMPLATE_INDEX_TYPE uint8_t
+#define _ZP_HASHSET_TEMPLATE_INITIAL_CAPACITY 2
+#define _ZP_HASHSET_TEMPLATE_KEY_HASH_FN u32_hash
+#define _ZP_HASHSET_TEMPLATE_KEY_EQ_FN u32_eq
+#include "zenoh-pico/collections/hashset_template.h"
 
 // ── Instantiate with a custom REALLOC_FN to exercise the in-place grow path ──
 
-#define _ZP_HASHMAP_TEMPLATE_KEY_TYPE uint32_t
-#define _ZP_HASHMAP_TEMPLATE_NAME reallocset
-#define _ZP_HASHMAP_TEMPLATE_INITIAL_CAPACITY 2
-#define _ZP_HASHMAP_TEMPLATE_KEY_HASH_FN u32_hash
-#define _ZP_HASHMAP_TEMPLATE_KEY_EQ_FN u32_eq
-#define _ZP_HASHMAP_TEMPLATE_REALLOC_FN(ptr, bytes) realloc(ptr, bytes)
-#include "zenoh-pico/collections/hashmap_template.h"
+#define _ZP_HASHSET_TEMPLATE_KEY_TYPE uint32_t
+#define _ZP_HASHSET_TEMPLATE_NAME reallocset
+#define _ZP_HASHSET_TEMPLATE_INITIAL_CAPACITY 2
+#define _ZP_HASHSET_TEMPLATE_KEY_HASH_FN u32_hash
+#define _ZP_HASHSET_TEMPLATE_KEY_EQ_FN u32_eq
+#define _ZP_HASHSET_TEMPLATE_REALLOC_FN(ptr, bytes) realloc(ptr, bytes)
+#include "zenoh-pico/collections/hashset_template.h"
 
 // ── Instantiate a set of owned_str_t to exercise the key destroy/move path ───
 
@@ -88,14 +87,14 @@ static inline void owned_str_move(owned_str_t *dst, owned_str_t *src) {
     src->str = NULL;
 }
 
-#define _ZP_HASHMAP_TEMPLATE_KEY_TYPE owned_str_t
-#define _ZP_HASHMAP_TEMPLATE_NAME strset
-#define _ZP_HASHMAP_TEMPLATE_INITIAL_CAPACITY 2
-#define _ZP_HASHMAP_TEMPLATE_KEY_HASH_FN owned_str_hash
-#define _ZP_HASHMAP_TEMPLATE_KEY_EQ_FN owned_str_eq
-#define _ZP_HASHMAP_TEMPLATE_KEY_MOVE_FN(d, s) owned_str_move(d, s)
-#define _ZP_HASHMAP_TEMPLATE_KEY_DESTROY_FN(p) free((p)->str)
-#include "zenoh-pico/collections/hashmap_template.h"
+#define _ZP_HASHSET_TEMPLATE_KEY_TYPE owned_str_t
+#define _ZP_HASHSET_TEMPLATE_NAME strset
+#define _ZP_HASHSET_TEMPLATE_INITIAL_CAPACITY 2
+#define _ZP_HASHSET_TEMPLATE_KEY_HASH_FN owned_str_hash
+#define _ZP_HASHSET_TEMPLATE_KEY_EQ_FN owned_str_eq
+#define _ZP_HASHSET_TEMPLATE_KEY_MOVE_FN(d, s) owned_str_move(d, s)
+#define _ZP_HASHSET_TEMPLATE_KEY_DESTROY_FN(p) free((p)->str)
+#include "zenoh-pico/collections/hashset_template.h"
 
 // ── Tests ─────────────────────────────────────────────────────────────────────
