+++ tests/z_static_hashset_template_test.c
@@ -25,19 +25,18 @@
 
 // ── Instantiate a uint32_t set, 12 buckets, capacity 12 ──────────────────────
 //
-// A set is generated by leaving _ZP_STATIC_HASHMAP_TEMPLATE_VAL_TYPE undefined.
 // Identity hash: bucket = key % capacity. Using the identity makes collisions
 // fully predictable, so the collision-chain tests below can deliberately place
 // several keys into the same bucket.
 static inline size_t u32_hash(const uint32_t *k) { return (size_t)(*k); }
 static inline bool u32_eq(const uint32_t *a, const uint32_t *b) { return *a == *b; }
 
-#define _ZP_STATIC_HASHMAP_TEMPLATE_KEY_TYPE uint32_t
-#define _ZP_STATIC_HASHMAP_TEMPLATE_NAME u32set
-#define _ZP_STATIC_HASHMAP_TEMPLATE_CAPACITY 12
-#define _ZP_STATIC_HASHMAP_TEMPLATE_KEY_HASH_FN u32_hash
-#define _ZP_STATIC_HASHMAP_TEMPLATE_KEY_EQ_FN u32_eq
-#include "zenoh-pico/collections/static_hashmap_template.h"
+#define _ZP_STATIC_HASHSET_TEMPLATE_KEY_TYPE uint32_t
+#define _ZP_STATIC_HASHSET_TEMPLATE_NAME u32set
+#define _ZP_STATIC_HASHSET_TEMPLATE_CAPACITY 12
+#define _ZP_STATIC_HASHSET_TEMPLATE_KEY_HASH_FN u32_hash
+#define _ZP_STATIC_HASHSET_TEMPLATE_KEY_EQ_FN u32_eq
+#include "zenoh-pico/collections/static_hashset_template.h"
 
 // ── Instantiate a set of owned_str_t to exercise the key destroy callback ────
 
@@ -63,13 +62,13 @@ static inline size_t owned_str_hash(const owned_str_t *k) {
 }
 static inline bool owned_str_eq(const owned_str_t *a, const owned_str_t *b) { return strcmp(a->str, b->str) == 0; }
 
-#define _ZP_STATIC_HASHMAP_TEMPLATE_KEY_TYPE owned_str_t
-#define _ZP_STATIC_HASHMAP_TEMPLATE_NAME strset
-#define _ZP_STATIC_HASHMAP_TEMPLATE_CAPACITY 8
-#define _ZP_STATIC_HASHMAP_TEMPLATE_KEY_HASH_FN owned_str_hash
-#define _ZP_STATIC_HASHMAP_TEMPLATE_KEY_EQ_FN owned_str_eq
-#define _ZP_STATIC_HASHMAP_TEMPLATE_KEY_DESTROY_FN(p) free((p)->str)
-#include "zenoh-pico/collections/static_hashmap_template.h"
+#define _ZP_STATIC_HASHSET_TEMPLATE_KEY_TYPE owned_str_t
+#define _ZP_STATIC_HASHSET_TEMPLATE_NAME strset
+#define _ZP_STATIC_HASHSET_TEMPLATE_CAPACITY 8
+#define _ZP_STATIC_HASHSET_TEMPLATE_KEY_HASH_FN owned_str_hash
+#define _ZP_STATIC_HASHSET_TEMPLATE_KEY_EQ_FN owned_str_eq
+#define _ZP_STATIC_HASHSET_TEMPLATE_KEY_DESTROY_FN(p) free((p)->str)
+#include "zenoh-pico/collections/static_hashset_template.h"
 
 // ── Tests ─────────────────────────────────────────────────────────────────────
